+++ CHANGES.rst
@@ -5,8 +5,12 @@ Changelog
 3.0.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Add `design.plone.contenttypes.behavior.argomenti_document` behavior to **Document**.
+  [cekk]
+- *correlato_in_evidenza* field now return also icon value in restapi calls.
+  [cekk]
+- Add leadimage to **CartellaModulistica**.
+  [cekk]
 
 3.0.3 (2021-05-20)
 ------------------